refactor: util package created and moved util classes
