+++ src/util/Status.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 public enum Status {
     NEW,
