+++ src/util/Type.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 public enum Type {
     TASK,
